bugfixes in sign up

+++ src/main/java/com/piotics/service/InvitationService.java
@@ -3,7 +3,6 @@ package com.piotics.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.piotics.common.TokenType;
 import com.piotics.model.Invitation;
 import com.piotics.model.Token;
 import com.piotics.repository.InvitationMongoRepository;
@@ -11,39 +10,27 @@ import com.piotics.repository.InvitationMongoRepository;
 @Service
 public class InvitationService {
 
-	@Autowired
-	InvitationMongoRepository invitationMongoRepository;
-	
-	@Autowired
-	TokenService tokenService;
-	
-	public Invitation save(Invitation invitation) {
-		
-		return invitationMongoRepository.save(invitation);
-	}
+    @Autowired
+    InvitationMongoRepository invitationMongoRepository;
 
-	public boolean isInvited(String username, Token token) {
+    @Autowired
+    TokenService tokenService;
 
-		Token dbToken = tokenService.getTokenFromDB(username);
+    public Invitation save(Invitation invitation) {
 
-		if (dbToken != null) {
+        return invitationMongoRepository.save(invitation);
+    }
 
-			if (tokenService.isTokenValid(dbToken)) {
+    public boolean isInvited(String username, Token token) {
 
-				if (dbToken.getTokenType().equals(TokenType.INVITATION)) {
+        Token dbToken = tokenService.getTokenFromDBWithTokenType(username,token.getTokenType());
 
-					return true;
-				} else {
-					return false;
-				}
+        if (dbToken != null && tokenService.isTokenValid(dbToken)) {
+            return true;
+        } else {
+            tokenService.deleteInviteToken(username, dbToken);
+            return false;
+        }
 
-			} else {
-				tokenService.deleteInviteToken(username, dbToken);
-				return false;
-			}
-		} else {
-
-			return false;
-		}
-	}
+    }
 }
